+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-### 2025-10-21 22:29:24 - f8d6760 [LOCAL]
+### 2025-10-29 23:02:32 - 36f9605 [LOCAL]
+
+feat(ui): enhance Markdown editor and renderer functionality
+
+- Improved text formatting options in MarkdownEditor
+- Added support for custom date formatting in date utility
+- Updated MarkdownRenderer to handle new formatting features### 2025-10-21 22:29:24 - f8d6760 [LOCAL]
 
 ### 2025-10-19 14:08:59 - 7c9ed27 [PUSHED]